Remove comparison warning.
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@8324 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Frontends/VirtualBox/src/VBoxConsoleWnd.cpp
@@ -1974,8 +1974,8 @@ bool VBoxConsoleWnd::toggleFullscreenMode (bool aOn, bool aSeamless)
         ULONG maxWidth  = guest.GetMaxGuestWidth();
         ULONG maxHeight = guest.GetMaxGuestHeight();
         if (aOn && (   (availBits < usedBits)
-                    || ((maxWidth != 0) && (maxWidth < screen.width()))
-                    || ((maxHeight != 0) && (maxHeight < screen.height()))
+                    || ((maxWidth != 0) && (maxWidth < (ULONG) screen.width()))
+                    || ((maxHeight != 0) && (maxHeight < (ULONG) screen.height()))
                    )
            )
         {
